backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
asgiref==3.11.0
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
    #   django-cors-headers
attrs==25.4.0
    # via
    #   service-identity
    #   twisted
autobahn==25.12.2
    # via daphne
automat==25.4.16
    # via twisted
cbor2==5.7.1
    # via autobahn
cffi==2.0.0
    # via
    #   autobahn
    #   cryptography
channels==4.3.2
    # via
    #   -r requirements.in
    #   channels-redis
channels-redis==4.3.0
    # via -r requirements.in
constantly==23.10.4
    # via twisted
cryptography==46.0.3
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==4.2.1
    # via -r requirements.in
django==5.2.9
    # via
    #   -r requirements.in
    #   channels
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==4.9.0
    # via -r requirements.in
django-environ==0.12.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via -r requirements.in
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.11
    # via
    #   hyperlink
    #   twisted
incremental==24.11.0
    # via twisted
msgpack==1.1.2
    # via
    #   autobahn
    #   channels-redis
packaging==25.0
    # via incremental
psycopg2-binary==2.9.11
    # via -r requirements.in
py-ubjson==0.16.1
    # via autobahn
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.2
    # via service-identity
pycparser==2.23
    # via cffi
pyopenssl==25.3.0
    # via twisted
redis==7.1.0
    # via channels-redis
service-identity==24.2.0
    # via twisted
sqlparse==0.5.5
    # via django
twisted[tls]==25.5.0
    # via daphne
txaio==25.12.2
    # via autobahn
typing-extensions==4.15.0
    # via twisted
ujson==5.11.0
    # via autobahn
zope-interface==8.1.1
    # via twisted
